+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All significant changes to this project will be documented here.
 
 ## [Unreleased]
 
+## [1.3]
+
+Release date: 2020-10-30
+
 *   Added a `StringArray` spec subclass for representing arrays of variable-
     length strings.
 *   Added a check to enforce that `minimum <= maximum` for `BoundedArray`.
@@ -39,7 +43,8 @@ Release date: 2019-07-18
 
 *   Initial release.
 
-[Unreleased]: https://github.com/deepmind/dm_env/compare/v1.2...HEAD
+[Unreleased]: https://github.com/deepmind/dm_env/compare/v1.3...HEAD
+[1.3]: https://github.com/deepmind/dm_env/compare/v1.2...v1.3
 [1.2]: https://github.com/deepmind/dm_env/compare/v1.1...v1.2
 [1.1]: https://github.com/deepmind/dm_env/compare/v1.0...v1.1
 [1.0]: https://github.com/deepmind/dm_env/releases/tag/v1.0